requirements.txt
opencv-python
pytesseract
pyttsx3
google-generativeai
Spire.Presentation
python-pptx